% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.norm.ui.test.R
\name{pca.norm.ui.test}
\alias{pca.norm.ui.test}
\title{This is for SixSigma > ss.study.ca : RcmdrPlugin.PCA}
\usage{
pca.norm.ui.test()
}
\value{
ss.study.ca result
}
\description{
This is for SixSigma > ss.study.ca : RcmdrPlugin.PCA
}
